Fixes issue with disabled tenants
The program would bomb with a stacktrace if it encountered any project
which was disabled. This is because by default it would expect that all
tenants/projects are enabled which is not always the case as was seen
few days back. This patch fixes this issue and ensure stats generation
complete even if there are OpenStack tenants in disabled state.